telegram-bot/requirements.txt
python-telegram-bot==20.7
instagrapi==1.17.0
python-dotenv==1.0.0
schedule==1.2.1
pymongo==4.6.1
python-jose==3.3.0
requests==2.31.0
aiohttp==3.9.1
pytz==2024.1
